backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart
google-generativeai>=0.8.0
sqlalchemy>=2.0.27
alembic>=1.13.1
asyncpg>=0.29.0
redis>=5.0.1
celery>=5.3.6
pydantic>=2.6.1
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
pymupdf==1.23.21
pillow>=10.2.0
tenacity>=8.2.3
aiofiles>=23.2.1
httpx>=0.27.0
